feat: add tweening functionality with easing methods and manager
