+++ .gitignore
@@ -3,3 +3,4 @@
 .vscode/
 node_modules
 npm-debug.log
+dist/
